@@ -1,25 +1,20 @@
 module github.com/quillpen
 
-go 1.18
+go 1.21.4
 
 require (
 	github.com/gocql/gocql v1.2.0
-	github.com/gorilla/csrf v1.7.1
 	github.com/gorilla/handlers v1.5.1
 	github.com/gorilla/mux v1.8.0
 	github.com/gorilla/sessions v1.2.1
-	golang.org/x/crypto v0.0.0-20220722155217-630584e8d5aa
+	github.com/gorilla/websocket v1.5.0
+	golang.org/x/crypto v0.4.0
 )
 
 require (
 	github.com/felixge/httpsnoop v1.0.1 // indirect
-	github.com/gobwas/httphead v0.1.0 // indirect
-	github.com/gobwas/pool v0.2.1 // indirect
-	github.com/gobwas/ws v1.2.1 // indirect
 	github.com/golang/snappy v0.0.3 // indirect
 	github.com/gorilla/securecookie v1.1.1 // indirect
-	github.com/gorilla/websocket v1.5.0 // indirect
 	github.com/hailocab/go-hostpool v0.0.0-20160125115350-e80d13ce29ed // indirect
-	github.com/pkg/errors v0.9.1 // indirect
 	gopkg.in/inf.v0 v0.9.1 // indirect
 )
